remove irismod/token dependency
