+++ go.mod
@@ -3,10 +3,10 @@ module github.com/irismod/service
 go 1.13
 
 require (
+	github.com/99designs/keyring v1.1.4 // indirect
 	github.com/cosmos/cosmos-sdk v0.38.2
+	github.com/golang/mock v1.4.3 // indirect
 	github.com/gorilla/mux v1.7.4
-	github.com/irismod/token v0.0.0-20200416081822-7ea072ec3f09
-	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
 	github.com/spf13/cobra v0.0.7
 	github.com/spf13/pflag v1.0.5
 	github.com/spf13/viper v1.6.2
@@ -14,4 +14,6 @@ require (
 	github.com/tendermint/tendermint v0.33.2
 	github.com/tendermint/tm-db v0.5.1
 	github.com/xeipuuv/gojsonschema v1.2.0
+	golang.org/x/net v0.0.0-20190827160401-ba9fcec4b297 // indirect
+	google.golang.org/protobuf v1.20.1 // indirect
 )
