+++ simapp/app.go
@@ -234,7 +234,7 @@ func NewSimApp(
 	)
 
 	app.ServiceKeeper = service.NewKeeper(
-		app.cdc, keys[service.StoreKey], app.SupplyKeeper, MockTokenKeeper{}, app.subspaces[service.ModuleName],
+		app.cdc, keys[service.StoreKey], app.SupplyKeeper, service.MockTokenKeeper{}, app.subspaces[service.ModuleName],
 	)
 
 	// NOTE: Any module instantiated in the module manager that is later modified
